Port to OTAPI3/.NET6

+++ FakeProvider/TileProvider/Tiles/StructTile.cs
@@ -10,13 +10,13 @@ namespace FakeProvider
     [StructLayout(LayoutKind.Sequential, Size = 13, Pack = 1)]
     public struct StructTile
     {
-        public byte wall;
+        public ushort wall;
         public byte liquid;
         public byte bTileHeader;
         public byte bTileHeader2;
         public byte bTileHeader3;
         public ushort type;
-        public short sTileHeader;
+        public ushort sTileHeader;
         public short frameX;
         public short frameY;
     }
